pyproject.toml
[project]
name = "yoga_app"
version = "0.1.0"
description = "A simple app"
authors = [
    {name = "Poseagile",email = "dmitrii.druzhbin@le-campus-numerique.fr"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "argparse (>=1.4.0,<2.0.0)",
    "numpy (==1.26.4)",
    "opencv-python (==4.11.0.86)",
    "tqdm (>=4.67.1,<5.0.0)",
    "mediapipe (>=0.10.21,<0.11.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "streamlit-webrtc (>=0.63.3,<0.64.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
